move runv_[linux|darwin].go to lib/utils
