+++ lib/utils/mount_darwin.go
@@ -1,14 +1,14 @@
-package main
+package utils
 
 import (
 	"syscall"
 )
 
-func mount(src, dst string) error {
+func Mount(src, dst string) error {
 	//return syscall.Symlink(src, dst)
 	return syscall.Link(src, dst)
 }
 
-func umount(path string) {
+func Umount(path string) {
 	syscall.Unlink(path)
 }
