+++ lib/utils/mount_linux.go
@@ -1,11 +1,11 @@
-package main
+package utils
 
 import (
 	"os"
 	"syscall"
 )
 
-func mount(src, dst string) error {
+func Mount(src, dst string) error {
 	if _, err := os.Stat(dst); os.IsNotExist(err) {
 		os.MkdirAll(dst, 0755)
 	}
@@ -13,6 +13,6 @@ func mount(src, dst string) error {
 	return syscall.Mount(src, dst, "", syscall.MS_BIND, "")
 }
 
-func umount(root string) {
+func Umount(root string) {
 	syscall.Unmount(root, syscall.MNT_DETACH)
 }
